REST API in process v 1.1

+++ src/main/java/ru/alexraydev/repository/uservote/UserVoteRepositoryImpl.java
@@ -2,12 +2,16 @@ package ru.alexraydev.repository.uservote;
 
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
+import ru.alexraydev.model.User;
 import ru.alexraydev.model.UserVote;
 import ru.alexraydev.repository.GenericRepository;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.util.List;
 
 @Repository
@@ -19,8 +23,13 @@ public class UserVoteRepositoryImpl implements UserVoteRepository{
 
     @Override
     @Transactional
-    public UserVote save(UserVote userVote) {
-        if (userVote.getId() == null) {
+    public UserVote save(UserVote userVote, int userId) {
+        if (!(userVote.getId() == null) && getById(userVote.getId(), userId) == null) {
+            return null;
+        }
+        userVote.setUser(em.getReference(User.class, userId));
+        if (userVote.getId() == null)
+        {
             em.persist(userVote);
             return userVote;
         }
@@ -30,21 +39,39 @@ public class UserVoteRepositoryImpl implements UserVoteRepository{
     }
 
     @Override
-    public UserVote getById(int id) {
-        return em.find(UserVote.class, id);
+    public UserVote getTodaysVote(int userId) {
+        return (UserVote)em.createQuery("SELECT uv FROM UserVote uv WHERE uv.user.id=:userId AND uv.date=:date")
+                .setParameter("userId", userId)
+                .setParameter("date", LocalDate.now())
+                .getSingleResult();
+    }
+
+    /*@Override
+    public UserVote getByDate(int userId, LocalDate date) {
+        UserVote userVote = (UserVote)em.createQuery("SELECT uv FROM UserVote uv WHERE uv.user.id=:userId AND uv.date=:date")
+                .setParameter("userId", userId)
+                .setParameter("date", date)
+                .getSingleResult();
+        return userVote;
+    }*/
+
+    @Override
+    public UserVote getById(int id, int userId) {
+        UserVote userVote = em.find(UserVote.class, id);
+        return userVote != null && userVote.getUser().getId() == userId ? userVote : null;
     }
 
     @Override
     @Transactional
-    public boolean delete(int id) {
-        Query query = em.createQuery("DELETE FROM UserVote uv WHERE uv.id=:id");
-        return query.setParameter("id", id).executeUpdate() != 0;
+    public boolean delete(int id, int userId) {
+        Query query = em.createQuery("DELETE FROM UserVote uv WHERE uv.id=:id AND uv.user.id=:userId");
+        return query.setParameter("id", id).setParameter("userId", userId).executeUpdate() != 0;
     }
 
     @SuppressWarnings("unchecked")
     @Override
-    public List<UserVote> getAll() {
-        Query query = em.createQuery("SELECT uv FROM UserVote uv");
-        return query.getResultList();
+    public List<UserVote> getAll(int userId) {
+        Query query = em.createQuery("SELECT uv FROM UserVote uv WHERE uv.user.id=:userId");
+        return query.setParameter("userId", userId).getResultList();
     }
 }
